refactor: split routes/index into routes and utils
