+++ webui/react/src/routes/routes.ts
@@ -1,6 +1,3 @@
-import { pathToRegexp } from 'path-to-regexp';
-import { RouteProps } from 'react-router';
-
 import Cluster from 'pages/Cluster';
 import Dashboard from 'pages/Dashboard';
 import ExperimentDetails from 'pages/ExperimentDetails';
@@ -12,26 +9,8 @@ import TaskList from 'pages/TaskList';
 import TaskLogs from 'pages/TaskLogs';
 import TrialDetails from 'pages/TrialDetails';
 import TrialLogs from 'pages/TrialLogs';
-import history from 'routes/history';
-import { isFullPath, parseUrl } from 'routes/utils';
-import { clone } from 'utils/data';
 
-/*
- * Router Configuration
- * If the component is not defined, the route is assumed to be an external route,
- * meaning React will attempt to load the path outside of the internal routing
- * mechanism.
- */
-export interface RouteConfig extends RouteProps {
-  id: string;
-  icon?: string;
-  path: string;
-  popout?: boolean;
-  redirect?: string;
-  suffixIcon?: string;
-  title?: string;
-  needAuth?: boolean;
-}
+import { RouteConfig } from './types';
 
 const dashboardRoute = {
   component: Dashboard,
@@ -143,42 +122,3 @@ export const appRoutes: RouteConfig[] = [
     redirect: defaultAppRoute.path,
   },
 ];
-
-// Add pages we don't want to expose to the public yet.
-if (process.env.IS_DEV) {
-  // appRoutes.push();
-}
-
-// Is the path going to be served from the same host?
-const isDetRoute = (url: string): boolean => {
-  if (!isFullPath(url)) return true;
-  if (process.env.IS_DEV) {
-    // dev live is served on a different port
-    return parseUrl(url).hostname === window.location.hostname;
-  }
-  return parseUrl(url).host === window.location.host;
-};
-
-const isReactRoute = (url: string): boolean => {
-  if (!isDetRoute(url)) return false;
-
-  // Check to see if the path matches any of the defined app routes.
-  const pathname = parseUrl(url).pathname;
-  return !!appRoutes
-    .filter(route => route.path !== '*')
-    .find(route => {
-      return route.exact ? pathname === route.path : !!pathToRegexp(route.path).exec(pathname);
-    });
-};
-
-export const routeToExternalUrl = (path: string): void => {
-  window.location.assign(path);
-};
-
-export const routeAll = (path: string): void => {
-  if (!isReactRoute(path)) {
-    routeToExternalUrl(path);
-  } else {
-    history.push(path, { loginRedirect: clone(window.location) });
-  }
-};
